services/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "clara-api"
version = "0.1.0"
description = "CLARA Backend API P0 skeleton"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "httpx>=0.27.2",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "psycopg[binary]>=3.2.3",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.13.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clara_api"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = true
